--- bundles/org.openhab.binding.solaredge/src/main/java/org/openhab/binding/solaredge/internal/handler/GenericSolarEdgeHandler.java
@@ -1,50 +1,0 @@
-/**
- * Copyright (c) 2010-2023 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.solaredge.internal.handler;
-
-import java.util.List;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.jetty.client.HttpClient;
-import org.openhab.core.thing.Channel;
-import org.openhab.core.thing.ChannelGroupUID;
-import org.openhab.core.thing.ChannelUID;
-import org.openhab.core.thing.Thing;
-import org.openhab.core.thing.ThingUID;
-
-/**
- * generic thing handler for solaredge
- *
- * @author Alexander Friese - initial contribution
- */
-@NonNullByDefault
-public class GenericSolarEdgeHandler extends SolarEdgeBaseHandler {
-
-    public GenericSolarEdgeHandler(Thing thing, HttpClient httpClient) {
-        super(thing, httpClient);
-    }
-
-    @Override
-    public List<Channel> getChannels() {
-        return getThing().getChannels();
-    }
-
-    @Override
-    public @Nullable Channel getChannel(String groupId, String channelId) {
-        ThingUID thingUID = this.getThing().getUID();
-        ChannelGroupUID channelGroupUID = new ChannelGroupUID(thingUID, groupId);
-        Channel channel = getThing().getChannel(new ChannelUID(channelGroupUID, channelId));
-        return channel;
-    }
-}
